Change Program Counter to 16 bit
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@638 75748be7-0387-4326-94c7-f41ce0883d2c

+++ sidplayfp/mos6510/cycle_based/mos6510c.h
@@ -99,7 +99,7 @@ protected:
     bool           flagB;
 
     /** Data regarding current instruction */
-    uint_least32_t Register_ProgramCounter;
+    uint_least16_t Register_ProgramCounter;
     uint_least16_t Cycle_EffectiveAddress;
     uint_least16_t Cycle_HighByteWrongEffectiveAddress;
     uint_least16_t Cycle_Pointer;
